+++ index.html
@@ -229,6 +229,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-<script src="main.js" type="module"></script>
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
